refactor(cli): `import` with subcommands for tables
Rather than having multiple table import commands, this consolidates them into one. You can either `import table` or `import bulk` to import a single table or a series of them under the same parent.
