+++ packages/cli/src/commands/index.ts
@@ -1,7 +1,6 @@
 import * as deployment from "./deployment.js";
 import * as importData from "./import-data.js";
 import * as importTable from "./import-table.js";
-import * as importTables from "./import-tables.js";
 import * as init from "./init.js";
 import * as login from "./login.js";
 import * as logout from "./logout.js";
@@ -17,7 +16,6 @@ export const commands = [
   init,
   importData,
   importTable,
-  importTables,
   login,
   logout,
   project,
